Modified-Atmosphere-Packaging-(MAP)-Equipment-Market.md
Modified Atmosphere Packaging (MAP) Equipment refers to packaging techniques that modify the atmosphere within a package to extend the shelf life of food products. It involves altering the proportion of gases in the package to slow down the natural degradation of perishable goods, maintaining product freshness and quality for an extended period.